requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.8
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.10.0
django-crispy-forms==1.12.0
django-filter==21.1
django-heroku==0.3.1
django-import-export==2.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-renderer-xlsx==0.4.1
et-xmlfile==1.1.0
gunicorn==20.1.0
httpie==2.5.0
idna==3.2
Markdown==3.3.4
MarkupPy==1.14
mysqlclient==2.0.3
numpy==1.21.2
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.8
pandas==1.3.3
Pillow==8.3.2
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.6
vncorenlp==1.0.3
whitenoise==5.3.0
xlrd==2.0.1
XlsxWriter==3.0.1
xlwt==1.3.0
